extract parseRequestObject out of NodeExecutionFunctions
