+++ packages/core/src/node-execute-functions.ts
@@ -14,7 +14,7 @@ import type {
 } from '@n8n/client-oauth2';
 import { ClientOAuth2 } from '@n8n/client-oauth2';
 import { Container } from '@n8n/di';
-import type { AxiosError, AxiosHeaders, AxiosRequestConfig, AxiosResponse } from 'axios';
+import type { AxiosError, AxiosRequestConfig, AxiosResponse } from 'axios';
 import axios from 'axios';
 import chardet from 'chardet';
 import crypto, { createHmac } from 'crypto';
@@ -33,7 +33,6 @@ import { extension, lookup } from 'mime-types';
 import type {
 	BinaryHelperFunctions,
 	FileSystemHelperFunctions,
-	GenericValue,
 	IAdditionalCredentialOptions,
 	IAllExecuteFunctions,
 	IBinaryData,
@@ -113,6 +112,7 @@ import { DataDeduplicationService } from './data-deduplication-service';
 // eslint-disable-next-line import/no-cycle
 import {
 	parseIncomingMessage,
+	parseRequestObject,
 	PollContext,
 	TriggerContext,
 } from './execution-engine/node-execution-context';
@@ -140,31 +140,6 @@ axios.interceptors.request.use((config) => {
 	return config;
 });
 
-const pushFormDataValue = (form: FormData, key: string, value: any) => {
-	if (value?.hasOwnProperty('value') && value.hasOwnProperty('options')) {
-		form.append(key, value.value, value.options);
-	} else {
-		form.append(key, value);
-	}
-};
-
-const createFormDataObject = (data: Record<string, unknown>) => {
-	const formData = new FormData();
-	const keys = Object.keys(data);
-	keys.forEach((key) => {
-		const formField = data[key];
-
-		if (formField instanceof Array) {
-			formField.forEach((item) => {
-				pushFormDataValue(formData, key, item);
-			});
-		} else {
-			pushFormDataValue(formData, key, formField);
-		}
-	});
-	return formData;
-};
-
 export const validateUrl = (url?: string): boolean => {
 	if (!url) return false;
 
@@ -184,29 +159,6 @@ function searchForHeader(config: AxiosRequestConfig, headerName: string) {
 	const headerNames = Object.keys(config.headers);
 	headerName = headerName.toLowerCase();
 	return headerNames.find((thisHeader) => thisHeader.toLowerCase() === headerName);
-}
-
-async function generateContentLengthHeader(config: AxiosRequestConfig) {
-	if (!(config.data instanceof FormData)) {
-		return;
-	}
-	try {
-		const length = await new Promise<number>((res, rej) => {
-			config.data.getLength((error: Error | null, length: number) => {
-				if (error) {
-					rej(error);
-					return;
-				}
-				res(length);
-			});
-		});
-		config.headers = {
-			...config.headers,
-			'content-length': length,
-		};
-	} catch (error) {
-		Container.get(Logger).error('Unable to calculate form data length', { error });
-	}
 }
 
 const getHostFromRequestObject = (
@@ -241,351 +193,6 @@ const getBeforeRedirectFn =
 			redirectedRequest.auth = `${axiosConfig.auth.username}:${axiosConfig.auth.password}`;
 		}
 	};
-
-// eslint-disable-next-line complexity
-export async function parseRequestObject(requestObject: IRequestOptions) {
-	// This function is a temporary implementation
-	// That translates all http requests done via
-	// the request library to axios directly
-	// We are not using n8n's interface as it would
-	// an unnecessary step, considering the `request`
-	// helper can be deprecated and removed.
-	const axiosConfig: AxiosRequestConfig = {};
-
-	if (requestObject.headers !== undefined) {
-		axiosConfig.headers = requestObject.headers as AxiosHeaders;
-	}
-
-	// Let's start parsing the hardest part, which is the request body.
-	// The process here is as following?
-	// - Check if we have a `content-type` header. If this was set,
-	//   we will follow
-	// - Check if the `form` property was set. If yes, then it's x-www-form-urlencoded
-	// - Check if the `formData` property exists. If yes, then it's multipart/form-data
-	// - Lastly, we should have a regular `body` that is probably a JSON.
-
-	const contentTypeHeaderKeyName =
-		axiosConfig.headers &&
-		Object.keys(axiosConfig.headers).find(
-			(headerName) => headerName.toLowerCase() === 'content-type',
-		);
-	const contentType =
-		contentTypeHeaderKeyName &&
-		(axiosConfig.headers?.[contentTypeHeaderKeyName] as string | undefined);
-	if (contentType === 'application/x-www-form-urlencoded' && requestObject.formData === undefined) {
-		// there are nodes incorrectly created, informing the content type header
-		// and also using formData. Request lib takes precedence for the formData.
-		// We will do the same.
-		// Merge body and form properties.
-		if (typeof requestObject.body === 'string') {
-			axiosConfig.data = requestObject.body;
-		} else {
-			const allData = Object.assign(requestObject.body || {}, requestObject.form || {}) as Record<
-				string,
-				string
-			>;
-			if (requestObject.useQuerystring === true) {
-				axiosConfig.data = stringify(allData, { arrayFormat: 'repeat' });
-			} else {
-				axiosConfig.data = stringify(allData);
-			}
-		}
-	} else if (contentType?.includes('multipart/form-data')) {
-		if (requestObject.formData !== undefined && requestObject.formData instanceof FormData) {
-			axiosConfig.data = requestObject.formData;
-		} else {
-			const allData: Partial<FormData> = {
-				...(requestObject.body as object | undefined),
-				...(requestObject.formData as object | undefined),
-			};
-
-			axiosConfig.data = createFormDataObject(allData);
-		}
-		// replace the existing header with a new one that
-		// contains the boundary property.
-		delete axiosConfig.headers?.[contentTypeHeaderKeyName!];
-		const headers = axiosConfig.data.getHeaders();
-		axiosConfig.headers = Object.assign(axiosConfig.headers || {}, headers);
-		await generateContentLengthHeader(axiosConfig);
-	} else {
-		// When using the `form` property it means the content should be x-www-form-urlencoded.
-		if (requestObject.form !== undefined && requestObject.body === undefined) {
-			// If we have only form
-			axiosConfig.data =
-				typeof requestObject.form === 'string'
-					? stringify(requestObject.form, { format: 'RFC3986' })
-					: stringify(requestObject.form).toString();
-			if (axiosConfig.headers !== undefined) {
-				const headerName = searchForHeader(axiosConfig, 'content-type');
-				if (headerName) {
-					delete axiosConfig.headers[headerName];
-				}
-				axiosConfig.headers['Content-Type'] = 'application/x-www-form-urlencoded';
-			} else {
-				axiosConfig.headers = {
-					'Content-Type': 'application/x-www-form-urlencoded',
-				};
-			}
-		} else if (requestObject.formData !== undefined) {
-			// remove any "content-type" that might exist.
-			if (axiosConfig.headers !== undefined) {
-				const headers = Object.keys(axiosConfig.headers);
-				headers.forEach((header) => {
-					if (header.toLowerCase() === 'content-type') {
-						delete axiosConfig.headers?.[header];
-					}
-				});
-			}
-
-			if (requestObject.formData instanceof FormData) {
-				axiosConfig.data = requestObject.formData;
-			} else {
-				axiosConfig.data = createFormDataObject(requestObject.formData as Record<string, unknown>);
-			}
-			// Mix in headers as FormData creates the boundary.
-			const headers = axiosConfig.data.getHeaders();
-			axiosConfig.headers = Object.assign(axiosConfig.headers || {}, headers);
-			await generateContentLengthHeader(axiosConfig);
-		} else if (requestObject.body !== undefined) {
-			// If we have body and possibly form
-			if (requestObject.form !== undefined && requestObject.body) {
-				// merge both objects when exist.
-				requestObject.body = Object.assign(requestObject.body, requestObject.form);
-			}
-			axiosConfig.data = requestObject.body as FormData | GenericValue | GenericValue[];
-		}
-	}
-
-	if (requestObject.uri !== undefined) {
-		axiosConfig.url = requestObject.uri?.toString();
-	}
-
-	if (requestObject.url !== undefined) {
-		axiosConfig.url = requestObject.url?.toString();
-	}
-
-	if (requestObject.baseURL !== undefined) {
-		axiosConfig.baseURL = requestObject.baseURL?.toString();
-	}
-
-	if (requestObject.method !== undefined) {
-		axiosConfig.method = requestObject.method;
-	}
-
-	if (requestObject.qs !== undefined && Object.keys(requestObject.qs as object).length > 0) {
-		axiosConfig.params = requestObject.qs;
-	}
-
-	function hasArrayFormatOptions(
-		arg: IRequestOptions,
-	): arg is Required<Pick<IRequestOptions, 'qsStringifyOptions'>> {
-		if (
-			typeof arg.qsStringifyOptions === 'object' &&
-			arg.qsStringifyOptions !== null &&
-			!Array.isArray(arg.qsStringifyOptions) &&
-			'arrayFormat' in arg.qsStringifyOptions
-		) {
-			return true;
-		}
-
-		return false;
-	}
-
-	if (
-		requestObject.useQuerystring === true ||
-		(hasArrayFormatOptions(requestObject) &&
-			requestObject.qsStringifyOptions.arrayFormat === 'repeat')
-	) {
-		axiosConfig.paramsSerializer = (params) => {
-			return stringify(params, { arrayFormat: 'repeat' });
-		};
-	} else if (requestObject.useQuerystring === false) {
-		axiosConfig.paramsSerializer = (params) => {
-			return stringify(params, { arrayFormat: 'indices' });
-		};
-	}
-
-	if (
-		hasArrayFormatOptions(requestObject) &&
-		requestObject.qsStringifyOptions.arrayFormat === 'brackets'
-	) {
-		axiosConfig.paramsSerializer = (params) => {
-			return stringify(params, { arrayFormat: 'brackets' });
-		};
-	}
-
-	if (requestObject.auth !== undefined) {
-		// Check support for sendImmediately
-		if (requestObject.auth.bearer !== undefined) {
-			axiosConfig.headers = Object.assign(axiosConfig.headers || {}, {
-				Authorization: `Bearer ${requestObject.auth.bearer}`,
-			});
-		} else {
-			const authObj = requestObject.auth;
-			// Request accepts both user/username and pass/password
-			axiosConfig.auth = {
-				username: (authObj.user || authObj.username) as string,
-				password: (authObj.password || authObj.pass) as string,
-			};
-		}
-	}
-
-	// Only set header if we have a body, otherwise it may fail
-	if (requestObject.json === true) {
-		// Add application/json headers - do not set charset as it breaks a lot of stuff
-		// only add if no other accept headers was sent.
-		const acceptHeaderExists =
-			axiosConfig.headers === undefined
-				? false
-				: Object.keys(axiosConfig.headers)
-						.map((headerKey) => headerKey.toLowerCase())
-						.includes('accept');
-		if (!acceptHeaderExists) {
-			axiosConfig.headers = Object.assign(axiosConfig.headers || {}, {
-				Accept: 'application/json',
-			});
-		}
-	}
-	if (requestObject.json === false || requestObject.json === undefined) {
-		// Prevent json parsing
-		axiosConfig.transformResponse = (res) => res;
-	}
-
-	// Axios will follow redirects by default, so we simply tell it otherwise if needed.
-	const { method } = requestObject;
-	if (
-		(requestObject.followRedirect !== false &&
-			(!method || method === 'GET' || method === 'HEAD')) ||
-		requestObject.followAllRedirects
-	) {
-		axiosConfig.maxRedirects = requestObject.maxRedirects;
-	} else {
-		axiosConfig.maxRedirects = 0;
-	}
-
-	const host = getHostFromRequestObject(requestObject);
-	const agentOptions: AgentOptions = { ...requestObject.agentOptions };
-	if (host) {
-		agentOptions.servername = host;
-	}
-	if (requestObject.rejectUnauthorized === false) {
-		agentOptions.rejectUnauthorized = false;
-		agentOptions.secureOptions = crypto.constants.SSL_OP_LEGACY_SERVER_CONNECT;
-	}
-
-	axiosConfig.httpsAgent = new Agent(agentOptions);
-
-	axiosConfig.beforeRedirect = getBeforeRedirectFn(agentOptions, axiosConfig);
-
-	if (requestObject.timeout !== undefined) {
-		axiosConfig.timeout = requestObject.timeout;
-	}
-
-	if (requestObject.proxy !== undefined) {
-		// try our best to parse the url provided.
-		if (typeof requestObject.proxy === 'string') {
-			try {
-				const url = new URL(requestObject.proxy);
-				const host = url.hostname.startsWith('[') ? url.hostname.slice(1, -1) : url.hostname;
-				axiosConfig.proxy = {
-					host,
-					port: parseInt(url.port, 10),
-					protocol: url.protocol,
-				};
-				if (!url.port) {
-					// Sets port to a default if not informed
-					if (url.protocol === 'http') {
-						axiosConfig.proxy.port = 80;
-					} else if (url.protocol === 'https') {
-						axiosConfig.proxy.port = 443;
-					}
-				}
-				if (url.username || url.password) {
-					axiosConfig.proxy.auth = {
-						username: url.username,
-						password: url.password,
-					};
-				}
-			} catch (error) {
-				// Not a valid URL. We will try to simply parse stuff
-				// such as user:pass@host:port without protocol (we'll assume http)
-				if (requestObject.proxy.includes('@')) {
-					const [userpass, hostport] = requestObject.proxy.split('@');
-					const [username, password] = userpass.split(':');
-					const [hostname, port] = hostport.split(':');
-					const host = hostname.startsWith('[') ? hostname.slice(1, -1) : hostname;
-					axiosConfig.proxy = {
-						host,
-						port: parseInt(port, 10),
-						protocol: 'http',
-						auth: {
-							username,
-							password,
-						},
-					};
-				} else if (requestObject.proxy.includes(':')) {
-					const [hostname, port] = requestObject.proxy.split(':');
-					axiosConfig.proxy = {
-						host: hostname,
-						port: parseInt(port, 10),
-						protocol: 'http',
-					};
-				} else {
-					axiosConfig.proxy = {
-						host: requestObject.proxy,
-						port: 80,
-						protocol: 'http',
-					};
-				}
-			}
-		} else {
-			axiosConfig.proxy = requestObject.proxy;
-		}
-	}
-
-	if (requestObject.useStream) {
-		axiosConfig.responseType = 'stream';
-	} else if (requestObject.encoding === null) {
-		// When downloading files, return an arrayBuffer.
-		axiosConfig.responseType = 'arraybuffer';
-	}
-
-	// If we don't set an accept header
-	// Axios forces "application/json, text/plan, */*"
-	// Which causes some nodes like NextCloud to break
-	// as the service returns XML unless requested otherwise.
-	const allHeaders = axiosConfig.headers ? Object.keys(axiosConfig.headers) : [];
-	if (!allHeaders.some((headerKey) => headerKey.toLowerCase() === 'accept')) {
-		axiosConfig.headers = Object.assign(axiosConfig.headers || {}, { accept: '*/*' });
-	}
-	if (
-		requestObject.json !== false &&
-		axiosConfig.data !== undefined &&
-		axiosConfig.data !== '' &&
-		!(axiosConfig.data instanceof Buffer) &&
-		!allHeaders.some((headerKey) => headerKey.toLowerCase() === 'content-type')
-	) {
-		// Use default header for application/json
-		// If we don't specify this here, axios will add
-		// application/json; charset=utf-8
-		// and this breaks a lot of stuff
-		axiosConfig.headers = Object.assign(axiosConfig.headers || {}, {
-			'content-type': 'application/json',
-		});
-	}
-
-	if (requestObject.simple === false) {
-		axiosConfig.validateStatus = () => true;
-	}
-
-	/**
-	 * Missing properties:
-	 * encoding (need testing)
-	 * gzip (ignored - default already works)
-	 * resolveWithFullResponse (implemented elsewhere)
-	 */
-	return axiosConfig;
-}
 
 function digestAuthAxiosConfig(
 	axiosConfig: AxiosRequestConfig,
